pyproject.toml
[project]
name = "chatbot-ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi>=0.121.1",
    "ipython>=9.6.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.10",
    "openai>=2.2.0",
    "pydantic-settings>=2.11.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
